backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
anthropic==0.40.0
pdfplumber==0.11.0
PyMuPDF==1.24.0
openpyxl==3.1.5
pydantic==2.9.0
python-dotenv==1.0.1
aiofiles==24.1.0
Pillow==10.4.0
asyncpg==0.30.0
boto3==1.35.0
sse-starlette==2.1.0
thefuzz==0.22.1
